update remove backup and add duplication for activity option

+++ api/src/models/TableHrbackups.js
@@ -1,6 +1,10 @@
-import { Op } from 'sequelize'
-import { USER_ROLE_ADMIN, USER_ROLE_SUPER_ADMIN, USER_ROLE_TEAM } from '../constants/roles'
-import { orderBy } from 'lodash'
+import { Op } from "sequelize";
+import {
+  USER_ROLE_ADMIN,
+  USER_ROLE_SUPER_ADMIN,
+  USER_ROLE_TEAM,
+} from "../constants/roles";
+import { orderBy } from "lodash";
 
 /**
  * Liste des juridctions auquels ont accès les utilisateur
@@ -14,20 +18,20 @@ export default (sequelizeInstance, Model) => {
    */
   Model.list = async (userId) => {
     const listAll = await Model.findAll({
-      attributes: ['id', 'label', ['updated_at', 'date'], 'jirs'],
+      attributes: ["id", "label", ["updated_at", "date"], "jirs"],
       include: [
         {
-          attributes: ['id'],
+          attributes: ["id"],
           model: Model.models.UserVentilations,
           where: {
             user_id: userId,
           },
         },
       ],
-      order: [['label', 'asc']],
+      order: [["label", "asc"]],
       raw: true,
-    })
-    const list = []
+    });
+    const list = [];
 
     for (let i = 0; i < listAll.length; i++) {
       if (await Model.models.TJ.isVisible(listAll[i].label)) {
@@ -36,29 +40,66 @@ export default (sequelizeInstance, Model) => {
           label: listAll[i].label,
           date: listAll[i].date,
           jirs: listAll[i].jirs,
-        })
+        });
       }
     }
-    return list
-  }
+    return list;
+  };
 
   /**
    * Suppression d'une juridiction
    * @param {*} backupId
    */
   Model.removeBackup = async (backupId) => {
-    await Model.destroy({
+    const backup = await Model.findOne({
       where: {
         id: backupId,
       },
-    })
+      raw: true,
+    });
 
-    await Model.models.HumanResources.destroy({
-      where: {
-        backup_id: backupId,
-      },
-    })
-  }
+    if (backup) {
+      // Recherche de HR
+      let hrList = await Model.models.HumanResources.findAll({
+        where: {
+          backup_id: backupId,
+        },
+        raw: true,
+      });
+      // Pour chaque HR
+      for (let x = 0; x < hrList.length; x++) {
+        const hrId = hrList[x].id;
+        // Recherche de situation pour chaque HR
+        await Model.models.HRSituations.destroy({
+          where: { human_id: hrId },
+        });
+      }
+
+      await Model.models.HumanResources.destroy({
+        where: {
+          backup_id: backupId,
+        },
+      });
+
+      await Model.models.Activities.destroy({
+        where: {
+          hr_backup_id: backupId,
+        },
+      });
+      
+      await Model.models.UserVentilations.destroy({
+        where: {
+          hr_backup_id: backupId,
+        },
+      });
+
+      await Model.destroy({
+        where: {
+          id: backupId,
+        },
+      });      
+    }
+  };
 
   /**
    * Copie d'un juridicition
@@ -66,22 +107,22 @@ export default (sequelizeInstance, Model) => {
    * @param {*} backupName
    * @returns
    */
-  Model.duplicateBackup = async (backupId, backupName) => {
+  Model.duplicateBackup = async (backupId, backupName,copyAct=false) => {
     const backup = await Model.findOne({
       where: {
         id: backupId,
       },
       raw: true,
-    })
+    });
 
     if (backup) {
       // Creation d'un BACKUP
-      delete backup.id
+      delete backup.id;
       const backupCreated = await Model.create({
         ...backup,
         label: backupName,
-      })
-      const newBackupId = backupCreated.dataValues.id
+      });
+      const newBackupId = backupCreated.dataValues.id;
 
       // Création d'un TJ ou CA IELST
       await models.TJ.create({
@@ -91,8 +132,8 @@ export default (sequelizeInstance, Model) => {
         longitude: 0,
         population: 0,
         enabled: true,
-        type: 'TGI'
-      })
+        type: "TGI",
+      });
 
       // Recherche de HR
       let hrList = await Model.models.HumanResources.findAll({
@@ -100,24 +141,24 @@ export default (sequelizeInstance, Model) => {
           backup_id: backupId,
         },
         raw: true,
-      })
+      });
 
-      hrList =orderBy(hrList, 'last_name')
+      hrList = orderBy(hrList, "last_name");
 
       for (let x = 0; x < hrList.length; x++) {
         // Pour chaque HR
-        const hrId = hrList[x].id
-        delete hrList[x].id
-        delete hrList[x]['first_name']
-        delete hrList[x]['last_name']
+        const hrId = hrList[x].id;
+        delete hrList[x].id;
+        delete hrList[x]["first_name"];
+        delete hrList[x]["last_name"];
 
         // Création de HR
         const newHR = await Model.models.HumanResources.create({
           ...hrList[x],
-          first_name: '',
+          first_name: "",
           last_name: x,
           backup_id: newBackupId,
-        })
+        });
 
         // Recherche de situation pour chaque HR
         const situationList = await Model.models.HRSituations.findAll({
@@ -125,75 +166,88 @@ export default (sequelizeInstance, Model) => {
             human_id: hrId,
           },
           raw: true,
-        })
+        });
 
         for (let y = 0; y < situationList.length; y++) {
           // Pour chaque situation
-          const situationId = situationList[y].id
-          delete situationList[y].id
+          const situationId = situationList[y].id;
+          delete situationList[y].id;
 
-          let category =  await models.HRCategories.findOne({
+          let category = await models.HRCategories.findOne({
             where: {
-              id: situationList[y]['category_id']
+              id: situationList[y]["category_id"],
             },
-          })
+          });
 
           // Renommer la HR si pas magistrat
-          if(category)
-            newHR.update({first_name: category.dataValues.label})
+          if (category) newHR.update({ first_name: category.dataValues.label });
 
           // Création d'une situation
           const newSituation = await Model.models.HRSituations.create({
             ...situationList[y],
             human_id: newHR.dataValues.id,
-          })
+          });
 
           // Recherche de ventilation dans chaque situation
           const ventilationList = await Model.models.HRActivities.findAll({
             where: {
-              hr_situation_id: situationId
+              hr_situation_id: situationId,
             },
             raw: true,
-          })
+          });
 
           for (let z = 0; z < ventilationList.length; z++) {
             // Pour chaque ventilation
-            delete ventilationList[z].id
+            delete ventilationList[z].id;
 
             // Création d'une ventilation
             await Model.models.HRActivities.create({
               ...ventilationList[z],
               hr_situation_id: newSituation.dataValues.id,
-            })
+            });
           }
-
         }
-
         // Recherche d'indispo
         const indispoList = await Model.models.HRIndisponibilities.findAll({
           where: {
             hr_id: hrId,
           },
           raw: true,
-        })
+        });
 
         for (let y = 0; y < indispoList.length; y++) {
           // Pour chaque indispo
-          delete indispoList[y].id
-          
+          delete indispoList[y].id;
+
           // Création d'une indispo
           await Model.models.HRIndisponibilities.create({
             ...indispoList[y],
             hr_id: newHR.dataValues.id,
-          })
+          });
         }
       }
 
-      return newBackupId
+      if(copyAct){
+        const activities = await Model.models.Activities.findAll({
+          where: {
+            hr_backup_id: backupId,
+          },
+          raw: true,
+        });
+        for (let w = 0; w < activities.length; w++) {
+          delete activities[w].id;
+          await Model.models.Activities.create({
+            ...activities[w],
+            hr_backup_id: newBackupId,
+          });
+        }
+      }
+
+      return newBackupId;
     } else {
-      return null
+      return null;
     }
-  }
+  };
 
   /**
    * Liste des juridictions
@@ -201,25 +255,25 @@ export default (sequelizeInstance, Model) => {
    */
   Model.getAll = async () => {
     const listAll = await Model.findAll({
-      attributes: ['id', 'label', ['created_at', 'date']],
-      order: [['label', 'asc']],
+      attributes: ["id", "label", ["created_at", "date"]],
+      order: [["label", "asc"]],
       raw: true,
-    })
-    const list = []
+    });
+    const list = [];
     for (let i = 0; i < listAll.length; i++) {
       if (await Model.models.TJ.isVisible(listAll[i].label)) {
         list.push({
           id: listAll[i].id,
           label: listAll[i].label,
           date: listAll[i].date,
-        })
+        });
       }
     }
     //console.log('\n\n\n\n\n\n\n\n\n\n\n\nLIST:', list)
     //list.map((elem) => console.log('id:', elem.id, ' | label:', elem.label))
     //console.log('\n\n\n\n\n\n\n\n\n\n\n\n')
-    return list
-  }
+    return list;
+  };
 
   /**
    * Control pour savoir si un utilisateur a accès à une juridiction id
@@ -232,10 +286,10 @@ export default (sequelizeInstance, Model) => {
       where: {
         id,
       },
-      attributes: ['id', 'label'],
+      attributes: ["id", "label"],
       include: [
         {
-          attributes: ['id'],
+          attributes: ["id"],
           model: Model.models.UserVentilations,
           required: true,
           where: {
@@ -244,10 +298,10 @@ export default (sequelizeInstance, Model) => {
         },
       ],
       raw: true,
-    })
+    });
 
-    return hr && (await Model.models.TJ.isVisible(hr.label)) ? true : false
-  }
+    return hr && (await Model.models.TJ.isVisible(hr.label)) ? true : false;
+  };
 
   /**
    * Création ou récupération d'une juridiction par son nom
@@ -256,22 +310,22 @@ export default (sequelizeInstance, Model) => {
    */
   Model.findOrCreateLabel = async (juridictionName) => {
     const find = await Model.findOne({
-      attributes: ['id'],
+      attributes: ["id"],
       where: {
         label: juridictionName,
       },
       raw: true,
-    })
+    });
 
     if (!find) {
       const newElement = await Model.create({
         label: juridictionName,
-      })
-      return newElement.dataValues.id
+      });
+      return newElement.dataValues.id;
     }
 
-    return find.id
-  }
+    return find.id;
+  };
 
   /**
    * Récupération d'une juridiction par son nom
@@ -280,19 +334,19 @@ export default (sequelizeInstance, Model) => {
    */
   Model.findByLabel = async (juridictionName) => {
     const find = await Model.findOne({
-      attributes: ['id'],
+      attributes: ["id"],
       where: {
         label: juridictionName,
       },
       raw: true,
-    })
+    });
 
     if (!find) {
-      return null
+      return null;
     }
 
-    return find.id
-  }
+    return find.id;
+  };
 
   /**
    * Récupération d'une juridiction par son id
@@ -301,17 +355,17 @@ export default (sequelizeInstance, Model) => {
    */
   Model.findById = async (id) => {
     const find = await Model.findOne({
-      attributes: ['label', 'jirs'],
+      attributes: ["label", "jirs"],
       where: { id },
       raw: true,
-    })
+    });
 
     if (!find) {
-      return null
+      return null;
     }
 
-    return find.label
-  }
+    return find.label;
+  };
 
   /**
    * Ajouter les accès aux membres de l'équipe à la juridiction lors de la création d'une juridiction
@@ -325,7 +379,7 @@ export default (sequelizeInstance, Model) => {
         },
         raw: true,
       })
-    ).map((s) => s.user_id)
+    ).map((s) => s.user_id);
 
     const users = await Model.models.Users.findAll({
       where: {
@@ -335,12 +389,15 @@ export default (sequelizeInstance, Model) => {
         },
       },
       raw: true,
-    })
+    });
 
     for (let i = 0; i < users.length; i++) {
-      await Model.models.UserVentilations.pushVentilation(users[i].id, juridicitionId)
+      await Model.models.UserVentilations.pushVentilation(
+        users[i].id,
+        juridicitionId
+      );
     }
-  }
+  };
 
-  return Model
-}
+  return Model;
+};
